feat: enhance terminal with dynamic imports and built-ins
