+++ __tests__/terminal.test.tsx
@@ -1,142 +1,81 @@
-jest.mock(
-  '@xterm/xterm',
-  () => ({
-    Terminal: jest.fn().mockImplementation(() => ({
-      open: jest.fn(),
-      focus: jest.fn(),
-      loadAddon: jest.fn(),
-      write: jest.fn(),
-      writeln: jest.fn(),
-      onData: jest.fn(),
-      onKey: jest.fn(),
-      dispose: jest.fn(),
-    })),
-  }),
-  { virtual: true }
-);
-jest.mock(
-  '@xterm/addon-fit',
-  () => ({
-    FitAddon: jest.fn().mockImplementation(() => ({ fit: jest.fn() })),
-  }),
-  { virtual: true }
-);
-jest.mock(
-  '@xterm/addon-search',
-  () => ({
-    SearchAddon: jest.fn().mockImplementation(() => ({
-      activate: jest.fn(),
-      dispose: jest.fn(),
-    })),
-  }),
-  { virtual: true }
-);
-jest.mock('@xterm/xterm/css/xterm.css', () => ({}), { virtual: true });
-jest.mock('react-ga4', () => ({ send: jest.fn(), event: jest.fn() }));
-
-import React, { createRef, act } from 'react';
-import { render } from '@testing-library/react';
-import Terminal from '../components/apps/terminal';
-
 jest.mock('react-ga4', () => ({ send: jest.fn(), event: jest.fn() }));
 jest.mock(
   'xterm',
-  () => ({
-    Terminal: class {
-      open() {}
-      write() {}
-      onData() {}
-    },
-  }),
-  { virtual: true },
-
+  () => {
+    const terminalMock = {
+      open: jest.fn(),
+      write: jest.fn(),
+      writeln: jest.fn(),
+      loadAddon: jest.fn(),
+      onData: jest.fn(),
+      onKey: jest.fn(),
+      clear: jest.fn(),
+      dispose: jest.fn(),
+    };
+    const Terminal = jest.fn(() => terminalMock);
+    return { Terminal, terminalMock, __esModule: true };
+  },
+  { virtual: true }
 );
 jest.mock(
   'xterm-addon-fit',
   () => ({
-    FitAddon: class {
-      fit() {}
-    },
+    FitAddon: jest.fn().mockImplementation(() => ({ fit: jest.fn() })),
+    __esModule: true,
   }),
-  { virtual: true },
-
+  { virtual: true }
 );
 jest.mock(
   'xterm-addon-search',
   () => ({
-    SearchAddon: class {
-      activate() {}
-    },
+    SearchAddon: jest.fn().mockImplementation(() => ({})),
+    __esModule: true,
   }),
-  { virtual: true },
+  { virtual: true }
 );
+jest.mock('xterm/css/xterm.css', () => ({}), { virtual: true });
 
+import React, { createRef, act } from 'react';
+import { render } from '@testing-library/react';
+import Terminal from '../components/apps/terminal';
+import { terminalMock } from 'xterm';
 
-describe.skip('Terminal component', () => {
+describe('Terminal component', () => {
   const addFolder = jest.fn();
   const openApp = jest.fn();
 
-  it('runs pwd command successfully', () => {
-    const ref = createRef();
-    render(<Terminal ref={ref} addFolder={addFolder} openApp={openApp} />);
+  const setup = async () => {
+    const ref = createRef<any>();
+    await act(async () => {
+      render(<Terminal ref={ref} addFolder={addFolder} openApp={openApp} />);
+    });
+    await act(async () => {});
+    return ref;
+  };
+
+  it('runs pwd command successfully', async () => {
+    const ref = await setup();
     act(() => {
       ref.current.runCommand('pwd');
     });
     expect(ref.current.getContent()).toContain('/home/alex');
   });
 
-  it('handles invalid cd command', () => {
-    const ref = createRef();
-    render(<Terminal ref={ref} addFolder={addFolder} openApp={openApp} />);
+  it('handles invalid cd command', async () => {
+    const ref = await setup();
     act(() => {
       ref.current.runCommand('cd nowhere');
     });
-    expect(ref.current.getContent()).toContain("bash: cd: nowhere: No such file or directory");
+    expect(ref.current.getContent()).toContain(
+      'bash: cd: nowhere: No such file or directory',
+    );
   });
 
-  it('supports history, clear, and help commands', () => {
-    const ref = createRef();
-    render(<Terminal ref={ref} addFolder={addFolder} openApp={openApp} />);
-    act(() => {
-      ref.current.runCommand('pwd');
-      ref.current.runCommand('history');
-    });
-    expect(ref.current.getContent()).toContain('pwd');
-    act(() => {
-      ref.current.runCommand('clear');
-    });
-    expect(ref.current.getContent()).toContain('pwd');
-    act(() => {
-      ref.current.runCommand('help');
-    });
-    expect(ref.current.getContent()).toContain('Available commands:');
-    expect(ref.current.getContent()).toContain('clear');
-    expect(ref.current.getContent()).toContain('help');
-  });
-
-  it('handles missing Worker gracefully', () => {
-    const ref = createRef();
-    const originalWorker = (global as any).Worker;
-    (global as any).Worker = undefined;
-    render(<Terminal ref={ref} addFolder={addFolder} openApp={openApp} />);
-    act(() => {
-      ref.current.runCommand('simulate');
-    });
-    expect(ref.current.getContent()).toContain('Web Workers are not supported');
-    (global as any).Worker = originalWorker;
-  });
-
-  it('navigates command history with arrow keys', () => {
-    const ref = createRef();
-    render(<Terminal ref={ref} addFolder={addFolder} openApp={openApp} />);
-    act(() => {
-      ref.current.runCommand('pwd');
-      ref.current.historyNav('up');
-    });
-    expect(ref.current.getCommand()).toBe('pwd');
-    act(() => {
-      ref.current.historyNav('down');
-    });
-    expect(ref.current.getCommand()).toBe('');
+  it('Ctrl+L clears terminal', async () => {
+    const ref = await setup();
+    const onKey = terminalMock.onKey.mock.calls[0][0];
+    const preventDefault = jest.fn();
+    onKey({ key: 'l', domEvent: { key: 'l', ctrlKey: true, preventDefault } });
+    expect(terminalMock.clear).toHaveBeenCalled();
   });
 });
