+++ components/apps/terminal/terminal.worker.js
